+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
+### Added
+- Modal error message when duplicate optional in the template is detected.
 
 
 ## [2.1.0] – 2025-10-02